sunset-mosaic/requirements.txt
numpy==2.0.2
pandas==2.2.3
pillow==11.2.1
tqdm==4.67.1